Finished Skeleton Code Phase 1
Phase 1 of creating the Skeleton Code for H is done. I have added all the appropriate and current code from subsystems such as drivetrain. We will have a meeting soon regarding what can be done for phase 2 of the skeleton code in preparation for the real deal in the coming weeks.