add notes on pathplannerlib changes
Some previously used pathplannerlib functions and classes are removed or moved. For the ones I could find another class of the same name, I have changed the import and made a note. For classes and functions without an equivalent, I have added a TODO and marked the offending import with a comment. Only Robot.java, Autonomous.java and CustomAutoBuilder.java are affected.